requirements.txt
# Tested with Python 3.7

# Prod:
Flask==1.1.1
python-dotenv
absl-py

# Dev:
pip
pipdeptree
setuptools
pycodestyle
flake8
flake8-quotes
flake8-isort
yapf
ipython
pyfakefs==3.6.1
pytest==5.3.0
pytest-env
mypy

# To install mypy in vscode, install the 'mypy' plugin and run these:
# sudo apt install python3.8-venv python3.8-dev
# python3.8 -m venv ~/.mypyls
# ~/.mypyls/bin/pip install wheel
# ~/.mypyls/bin/pip install "https://github.com/matangover/mypyls/archive/master.zip#egg=mypyls[patched-mypy]"
